changed isDomestic type to StudentType-string for easier search/filtering
